Refactor: Rename the project from POSQ to POUQ and refactor the related code
Refactor the project name and related code and rename POSQ to POUQ. Updated the project file, class name, variable name, and references in the document. New implementations of core modules such as bitmap, utils, quantizer, etc. have been added to optimize the performance of IVF index and quantizer. The test script and experimental code are also perfected to adapt to the new name.

The main changes include:
1. Rename all POSQ related identifiers to POUQ
2. Added bitmap module to implement bit operation function
3. Add the utils tool class to provide SIMD optimization calculations
4. Implement the core logic of quantizer
5. Optimize clustering and search performance of IVF indexes
6. Update test scripts and experimental code